etl-web/requirements.txt
本工程依赖以下包
werkzeug-0.11.3
jinja2-2.8
flask-0.10.1
visitor-0.1.2
dominate-2.1.17
Flask-Bootstrap-3.3.5.7
cx_Oracle
sqlalchemy
